tests: drivers: optee: workaround for ENOMEM
The optee suite was failing in test_suspend() after a call to
k_aligned_alloc() which returns -ENOMEM. The alignment requirement is
4096 bytes, which seems reasonable, but what is odd about that is
that the alignment fails in spite of

CONFIG_SRAM_SIZE=145131134582784
and
CONFIG_HEAP_MEM_POOL_SIZE=270044

So there is plenty of memory available for this test and alignment
considerations should not be an issue.

Removing `CONFIG_DEBUG=y` solved the problem for me.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>