samples: sensor: accel_trig: Fix harness configuration
Error message HARNESS:Console:no regex patterns configured
is shown when sample.sensor.accel_trig sample is run.
The error occurs because harness configuration is changed
(a fixture is added) in the test case. When the configuration
is changed in a test case, common configuration is not
inherited and required properties need to be redefined.
Add console harness configuration for the test case.

Signed-off-by: Michal Smola <michal.smola@nxp.com>