tests: Bluetooth: Tester: Add missing target latency and PHY
The struct bt_audio_codec_cfg recently got new fields that
previously defaulted to
BT_AUDIO_CODEC_CFG_TARGET_LATENCY_BALANCED and
BT_AUDIO_CODEC_CFG_TARGET_PHY_2M but was not set by the BTP
commands.

For now we restore the previous functionality by setting the
values to the same as we did before, until the BTP commands
are changed (if they are changed).

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>