secure_storage: its: use %l printf length modifier
Cast the UIDs to unsigned long when they are 32 bits.
This allows to use a single %l length modifier instead of the double
one for long long.

Certain printf implementations only support the former and not the
latter length modifier, so this has the advantage to work with them
now that the UIDs are 32-bit by default.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>