docs boards native_sim: Change title
Let's better refer to the board by the name "native_sim",
and leave the "native simulator" for the underlaying common
component which is shared with the nrf*bsim targets.

native_sim is the "plain vanilla" native simulator build with
some trivial HW models.
While the nrf*bsim targets use the native simulator (without those
trivial HW models), and expand it with the Nordic HW models.

Hopefully that helps people understand what is common and what
is not.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>