display: add PIXEL_FORMAT_AL_88
When rendering an overlay on top of something like a video stream on a
monochrome display, the only way to achieve this right now is by using
ARGB8888 and discarding two of the colors, which is not ideal.

This introduces the new AL_88 pixel format, which is much more efficient
for this usecase.

This change is somewhat inspired by LVGL, which also supports AL_88
natively.

Signed-off-by: Martin Stumpf <finomnis@gmail.com>