Bluetooth: ATT: Change LOG_ERR to DBG when failing to allocate buffer
For user-initiated requests like notifications, read and write
requests, there will be an error code, so there does not need to be
a generic LOG_ERR if these fail to be allocated.

However if the response or (indication) confirmation PDUs fail
to be allocated, these should still be logged as errors.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>